server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-chess==1.999
pydantic==2.8.2
python-dotenv==1.0.1
redis==5.0.7